Introduce Optional Workspaces
Prior to this commit any Workspaces declared by a Task are required to
be provided by a TaskRun or Pipeline/Run.

This commit introduces the concept of Optional Workspaces: a Task may
declare a Workspace as optional and a TaskRun may then omit that
Workspace at runtime. Pipelines and PipelineRuns may do the same.

A new variable is introduced for workspaces, "bound", which interpolates
to either "true" or "false" depending on whether a workspace binding is
included in a TaskRun or not.

Default TaskRun Workspaces are _not_ injected in place of an omitted
optional Workspace.